Refactor themes to be part of settings
- changed from having a themes section to keeping theme related files
under settings
