+++ lib/enums/settings/brightness_mode.dart
