+++ lib/extensions/theme_data_extensions.dart
@@ -1,5 +1,5 @@
-import 'package:crowdleague/enums/themes/theme_brightness.dart';
-import 'package:crowdleague/models/themes/theme_set.dart';
+import 'package:crowdleague/enums/settings/theme_brightness.dart';
+import 'package:crowdleague/models/settings/theme_set.dart';
 import 'package:flutter/material.dart';
 
 // static functions must be called on the extension name, ie. ThemeDataExt
