+++ lib/extensions/theme_mode_extensions.dart
@@ -1,10 +1,12 @@
-import 'package:crowdleague/enums/themes/brightness_mode.dart';
+import 'package:crowdleague/enums/settings/brightness_mode.dart';
 import 'package:flutter/material.dart';
 
 extension ThemeModeExt on ThemeMode {
   static ThemeMode from(BrightnessMode brightness) {
     return (brightness.isLight)
         ? ThemeMode.light
-        : (brightness.isDark) ? ThemeMode.dark : ThemeMode.system;
+        : (brightness.isDark)
+            ? ThemeMode.dark
+            : ThemeMode.system;
   }
 }
