+++ lib/models/app/app_state.dart
@@ -8,7 +8,6 @@ import 'package:built_value/serializer.dart';
 import 'package:crowdleague/enums/device/platform_type.dart';
 import 'package:crowdleague/enums/nav_bar_selection.dart';
 import 'package:crowdleague/models/app/problem.dart';
-import 'package:crowdleague/models/app/settings.dart';
 import 'package:crowdleague/models/app/system_info.dart';
 import 'package:crowdleague/models/auth/user.dart';
 import 'package:crowdleague/models/auth/vm_auth_page.dart';
@@ -19,6 +18,7 @@ import 'package:crowdleague/models/conversations/vm_conversation_summaries_page.
 import 'package:crowdleague/models/functions/processing_failure.dart';
 import 'package:crowdleague/models/navigation/route_info.dart';
 import 'package:crowdleague/models/profile/vm_profile_page.dart';
+import 'package:crowdleague/models/settings/settings.dart';
 import 'package:crowdleague/models/storage/upload_task.dart';
 import 'package:crowdleague/utils/serializers.dart';
