+++ lib/models/settings/theme_set.dart
@@ -4,8 +4,8 @@ import 'dart:convert';
 
 import 'package:built_value/built_value.dart';
 import 'package:built_value/serializer.dart';
-import 'package:crowdleague/enums/themes/theme_brightness.dart';
-import 'package:crowdleague/models/themes/theme_colors.dart';
+import 'package:crowdleague/enums/settings/theme_brightness.dart';
+import 'package:crowdleague/models/settings/theme_colors.dart';
 import 'package:crowdleague/utils/serializers.dart';
 import 'package:meta/meta.dart';
