+++ lib/widgets/crowd_league_app.dart
@@ -5,8 +5,8 @@ import 'package:crowdleague/actions/notifications/print_fcm_token.dart';
 import 'package:crowdleague/actions/notifications/request_fcm_permissions.dart';
 import 'package:crowdleague/extensions/extensions.dart';
 import 'package:crowdleague/models/app/app_state.dart';
-import 'package:crowdleague/models/app/settings.dart';
 import 'package:crowdleague/models/auth/user.dart';
+import 'package:crowdleague/models/settings/settings.dart';
 import 'package:crowdleague/utils/redux/navigation_info_recorder.dart';
 import 'package:crowdleague/utils/redux/services_bundle.dart';
 import 'package:crowdleague/utils/wrappers/firebase_wrapper.dart';
