+++ lib/widgets/more_options/more_options_page.dart
@@ -1,7 +1,7 @@
 import 'package:crowdleague/actions/auth/sign_out_user.dart';
-import 'package:crowdleague/actions/themes/store_theme_colors.dart';
+import 'package:crowdleague/actions/settings/store_theme_colors.dart';
 import 'package:crowdleague/extensions/extensions.dart';
-import 'package:crowdleague/models/themes/theme_colors.dart';
+import 'package:crowdleague/models/settings/theme_colors.dart';
 import 'package:crowdleague/widgets/more_options/dark_mode_toggle.dart';
 import 'package:flutter/material.dart';
